Integrate backend auth and registration in frontend
Replaces mock login/registration with real API calls using Axios, adds .env for API URL, and updates context and pages to handle authentication and registration with backend. Updates Header to use new roles and navigation logic, and improves UI for login and registration pages. Removes unused assets and updates favicon.

+++ src/main/java/com/greencycle/ecoswap/ecoswap/controller/OrdenController.java
@@ -102,7 +102,7 @@ public class OrdenController {
 
     // LISTAR MIS ORDENES (Para la Recicladora)
     @GetMapping("/mis-ordenes/{usuarioId}")
-    public List<Orden> listarPorUsuario(@PathVariable Long usuarioId) {
+    public List<Orden> listarPorUsuario(@PathVariable Integer usuarioId) {
         return ordenRepository.findByUsuarioId(usuarioId);
     }
 }
